updated types
