+++ types/auth.go
@@ -11,17 +11,3 @@ func NewAccount(address string) Account {
 		Address: address,
 	}
 }
-
-// TopAccount represents a cheqd top account
-type TopAccount struct {
-	Address string
-	Type    string
-}
-
-// TopAccount builds a new TopAccount instance
-func NewTopAccount(address, accountType string) TopAccount {
-	return TopAccount{
-		Address: address,
-		Type:    accountType,
-	}
-}
